requirements.txt
appdirs==1.4.3
boto==2.47.0
dj-database-url==0.4.1
Django==1.10.6
django-markdownx==2.0.21
django-storages==1.6.3
django-storages-redux==1.3.3
gunicorn==19.6.0
Markdown==2.6.8
olefile==0.44
packaging==16.8
Pillow==4.0.0
psycopg2==2.7.1
pyparsing==2.2.0
requests==2.9.1
six==1.10.0
whitenoise==3.3.0
